Main menu added with working play button. States coded ( not done ). Spawning in added. Menu was made using ROACT AND RODUX. Implemented states useing rodux.
I added a main menu using Roact + Rodux with just a play button that when pressed spawns the player, and updates the states in the server and client. The server handles the spawning and when the player dies the server sends a signal to change the state in the client to main menu, and there is a playing state too. The server and client stores and manages states using RODUX.